requirements.txt
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.28.0
cycler==0.12.1
easydict==1.9
filelock==3.13.1
hparams==0.3.0
idna==3.6
imageio==2.9.0
importlib-metadata==7.0.0
Jinja2==2.11.3
joblib==1.3.2
json5==0.9.4
jsonschema==3.2.0
kiwisolver==1.4.5
lit==17.0.6
Markdown==3.1.1
MarkupSafe==1.1.1
mat4py==0.4.2
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.1.2.30
pandas==2.0.3
Pillow==6.2.1
progress==1.5
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scikit-image==0.17.2
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.10.0
six==1.16.0
sympy==1.12
tabulate==0.8.7
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==2.0.1
torch-summary==1.4.5
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.0.3
tqdm==4.66.0
triton==2.0.0
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.25.9
xmltodict==0.12.0
zipp==3.17.0
